requirements.txt
sphinx
sphinx-collapse
sphinx-better-theme